+++ processor/src/test/fixtures/bad_input/com/example/injector/MissingPrimitiveDependencyModel.java
@@ -1,0 +1,31 @@
+package com.example.injector;
+
+import javax.annotation.Nullable;
+import sting.Fragment;
+import sting.Injectable;
+import sting.Injector;
+
+@Injector
+interface MissingPrimitiveDependencyModel
+{
+  MyModel1 getMyModel1();
+
+  @Injectable
+  class MyModel1
+  {
+    MyModel1( @Nullable String config )
+    {
+    }
+  }
+
+  @Fragment
+  interface MyFragment
+  {
+    // Nullable provides
+    @Nullable
+    default String provideConfig( int number )
+    {
+      return null;
+    }
+  }
+}